Update virtual-network-concepts.md - Include a disclaimer that IP assignments happen dynamically and may change depending upon circumstances
Working in CSS, I have observed that this is a point of confusion to many customers. Mainly because they wish to add DNS entries beforehand so are looking to predict the apim private vip.

Especially those who are migrating to STV2, they wish to have everything else ready to reduce downtime.